// frontend/src/client/types.gen.ts
export type AssistantMessage = {
    /**
     * Conversation Id
     */
    conversation_id: string;
    response_message: SimplifiedMessageCmd;
};

/**
 * Body_uploadFiles
 */
export type BodyUploadFiles = {
    /**
     * Uploadfile
     */
    uploadFile: Blob | File;
    /**
     * Gpt Ids
     */
    gpt_ids: Array<string>;
};

/**
 * ChatHistory
 */
export type ChatHistory = {
    /**
     * Custom Gpt Id
     */
    custom_gpt_id: string | null;
    /**
     * Messages
     */
    messages: Array<SimplifiedMessageQueries>;
};
// ...
export type RoleCmd = 'user' | 'assistant';

/**
 * RoleQuery
 */
export type RoleQuery = 'user' | 'assistant';

/**
 * SimplifiedMessageCmd
 */
export type SimplifiedMessageCmd = {
    role: RoleCmd;
    /**
     * Message
     */
    message: string;
};

/**
 * SimplifiedMessageQueries
 */
export type SimplifiedMessageQueries = {
    role: RoleQuery;
    /**
     * Message
     */
    message: string;
};